% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareReciprocalContrasts.R
\name{compare_reciprocal_contrasts}
\alias{compare_reciprocal_contrasts}
\title{Compare results from reciprocal contrasts}
\usage{
compare_reciprocal_contrasts(groups, folder)
}
\arguments{
\item{groups}{character vector specifying the names of the levels in the
factor used for the DESeq analysis by \code{\link[=run_DESeq_all_contrasts]{run_DESeq_all_contrasts()}}. i.e. if
\code{\link[=run_DESeq_all_contrasts]{run_DESeq_all_contrasts()}} had the parameters \verb{dds = dds, condition = 'set'},
then here you could use \code{groups = levels(dds$set)}}

\item{folder}{character string with the file path to the folder of results
from \code{\link[=run_DESeq_all_contrasts]{run_DESeq_all_contrasts()}}. Output figures and tables will also be
deposited here.}
}
\value{
named list of data.frames representing merged tables from each 'pair'
of contrasts. Also creates pdf and csv files in the specified folder.
}
\description{
This finds all the 'pairs' among the DG csvs created by
\code{\link[=run_DESeq_all_contrasts]{run_DESeq_all_contrasts()}} and looks for differences in the results found
when using each level as the reference level.
}
\section{Rationale}{


Let the pair of levels in question be called group1 and group2. Then the
genes found to be upregulated when using group1 as the reference should be
downregulated when group2 is the reference, and vice versa. Also the
magnitude of the log fold change for a gene should be nearly equal for both
sets of results, although with the opposite sign; thus the sum of the two
log fold changes for each gene is expected to be distributed around zero.
Furthermore, there should be a strong concordant set of significantly
up/down-regulated genes i.e. high degree of overlap between the genes seen
using group1 as the reference and those seen using group2. However,
\code{\link[=lfcShrink]{lfcShrink()}} can sometimes produce quite different results for a contrast
depending which level you use as reference, and from our observation is
then more likely to find upregulated genes than down-regulated genes, which
leads to low concordance between the results sets and sum(log fold changes)
density plots which are skewed/shifted to the right. This function produces
\enumerate{
\item simple Venn diagrams showing the overlap in results for the pairwise
comparisons, collected in one pdf named Venns.pdf;
}
\enumerate{
\item diagnostic plots to visualise how well the analysis worked, including
density plots of the sums of the log fold changes per gene, and stacked bar
plots summarising the Venn diagrams so that the strength of concordance
can be judged readily;
\item csv tables containing only the concordant genes for each pair of levels;
\item Finally, the overlap in the identities of significant genes between
different pairwise contrasts may be of interest. The function uses UpSetR
to plot and print the intersections between the concordant DEG sets from
each pairwise contrast.
}
}

