% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorCheck-package.R
\docType{package}
\name{mirrorCheck-package}
\alias{mirrorCheck}
\alias{mirrorCheck-package}
\title{mirrorCheck: Facilitator functions for getting and assessing DESeq2 lfcShrink results}
\description{
mirrorCheck is intended to facilitate use and understanding of the DESeq2 package especially when using lfcShrink with the apeglm algorithm. It includes two functions aimed at getting and assessing DESeq2 lfcShrink results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kescull/mirrorCheck}
  \item Report bugs at \url{https://github.com/kescull/mirrorCheck/issues}
}

}
\author{
\strong{Maintainer}: Kate Scull \email{kate.scull@monash.edu}

}
\keyword{internal}
