% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeqAllContrasts.R
\name{run_DESeq_all_contrasts}
\alias{run_DESeq_all_contrasts}
\title{Run DESeq for all contrasts}
\usage{
run_DESeq_all_contrasts(
  dds,
  folder,
  mode = "lfcShrink",
  condition,
  rowname2symbol = NULL,
  heatmap.annotation.col = NULL,
  p.cutoff = 0.1,
  fc.cutoff = 2,
  top.n = 30,
  only.first.iteration = F
)
}
\arguments{
\item{dds}{Valid \code{DESeqDataSet} object with a simple design e.g. \code{design = ~condition}. If you wish to use a filter such as \code{edgeR} \code{\link[=filterByExpr]{filterByExpr()}},
please do so prior to running this function. This package is intended to
help assess the usefulness of such filtering.}

\item{folder}{character string specifying the relative or absolute filepath
of the folder where you want the output files to go. This folder should
exist.}

\item{mode}{either \code{"noShrink"} or \code{"lfcShrink"}. Specifies whether to use
\code{\link[=results]{results()}} or \code{\link[=lfcShrink]{lfcShrink()}} (default is \code{"lfcShrink"}).}

\item{condition}{character string related to the design formula used when
creating the \code{DESeqDataSet}. This should be a \code{colData} variable in the
\code{DESeqDataSet}; e.g. for \code{design=~group}, here enter \code{condition="group"}.
If you use a more complex design, results are not guaranteed. See also
\code{only.first.iteration}.}

\item{rowname2symbol}{data.frame or tibble with 2 columns titled \code{"rownames"}
and \code{"g_symbol"}. \code{"rownames"} should match \code{"rownames(dds)"}. This is used
for labeling the Volcano plot and annotating the tables of differentially
expressed genes}

\item{heatmap.annotation.col}{character vector specifying which \code{colData}
variables to use as groups in the heatmap output. If NULL, the heatmap will
use the variable specified in \code{condition} (default NULL)}

\item{p.cutoff}{numeric. The significance cut-off for p-adjusted values,
equivalent to parameter alpha in \code{\link[DESeq2:results]{DESeq2::results()}}. (default 0.1)}

\item{fc.cutoff}{numeric. The log2 fold change significance cut-off; this is
used in the Volcano plots and when printing out significant genes. It is
NOT used to change the null hypothesis in \code{\link[DESeq2:results]{DESeq2::results()}}. (default 2)}

\item{top.n}{whole number. Number of most highly regulated genes to label
with gene symbols on Volcano plots. (default 30).}

\item{only.first.iteration}{If True, the function will process all the
contrasts using the first level as the reference level, then exit. This
should produce 1 set of csvs for contrast DGEs and 1 pdf with MA and
Volcano plots. It may be helpful for looking at more complex designs, so
long as \code{condition} can still be specified as a \code{colData} variable. But no
promises. (default False)}
}
\value{
None - but creates pdf and csv files in the specified folder.
}
\description{
Runs \code{\link[=DESeq]{DESeq()}} for a \code{DESeqDataSet} with a simple design. It uses each
possible reference level in turn, extracting all contrast results for each.
}
\section{Further details}{
 Output includes pdf files with DESeq2 MA plots,
volcano plots and heatmaps of significant genes for each contrast in one
pdf file per reference level, and also a csv of the significant genes for
each contrast (according to the user-specified adj-pvalue and lfc
thresholds). It is possible to run with more complex designs for the first
reference level using the \code{only.first.iteration} parameter, but this has
not been tested thoroughly and output should be interpreted with care.
}

