% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeqAllContrasts.R
\name{run_DESeq_all_contrasts}
\alias{run_DESeq_all_contrasts}
\title{Run DESeq for all contrasts}
\usage{
run_DESeq_all_contrasts(
  dds,
  folder,
  mode = "apeglm",
  condition,
  rowname2symbol = NULL,
  heatmap.annotation.col = NULL,
  p.cutoff = 0.1,
  fc.cutoff = 2,
  top.n = 30,
  useDingbats = F,
  print.all = F
)
}
\arguments{
\item{dds}{Valid \code{DESeqDataSet} object with a formula design e.g.
\code{design = ~condition}. Designs may be multi-factorial (e.g.
\code{design = ~batch + condition}), but may not include interaction terms. If
you wish to clean the data, such as applying a filter such as \code{edgeR}
\code{\link[=filterByExpr]{filterByExpr()}}, please do so prior to running this function. This package
is intended to help assess the usefulness of such cleaning steps.}

\item{folder}{character string specifying the relative or absolute filepath
of the folder where you want the output files to go. This folder should
exist.}

\item{mode}{one of \code{"none"}, \code{"apeglm"}, \code{"ashr"} or \code{"normal"}. Specifies
whether to apply LFC shrinkage and with which algorithm (default is
\code{"apeglm"}).}

\item{condition}{character string related to the design formula used when
creating the \code{DESeqDataSet}. This should be a \code{colData} variable in the
\code{DESeqDataSet}; e.g. for \code{design=~group}, enter \code{condition="group"} here.
\code{\link[=run_DESeq_all_contrasts]{run_DESeq_all_contrasts()}} will estimate dispersion when the first level
of the factor (e.g. the first level in 'group') is the reference level. You
can use a design with extra variables to account for additional variation,
e.g. for \code{design=~batch + group}, enter \code{condition="group"}: in this case,
batch effects will be accounted for and \code{\link[=run_DESeq_all_contrasts]{run_DESeq_all_contrasts()}} will
only extract results for contrasts between the levels in 'group'.}

\item{rowname2symbol}{data.frame or tibble with 2 columns titled \code{"rowname"}
and \code{"g_symbol"}. \code{"rowname"} should match \code{"rownames(dds)"}. This is used
for labeling the Volcano plot and annotating the tables of differentially
expressed genes}

\item{heatmap.annotation.col}{character vector specifying which \code{colData}
variables to use as groups in the heatmap output. If NULL, the heatmap will
use the variable specified in \code{condition} (default NULL)}

\item{p.cutoff}{numeric. The significance cut-off for p-adjusted values,
equivalent to parameter alpha in \code{\link[DESeq2:results]{DESeq2::results()}}. (default 0.1)}

\item{fc.cutoff}{numeric. The log2 fold change significance cut-off; this is
used in the Volcano plots and when printing out significant genes. It is
NOT used to change the null hypothesis in \code{\link[DESeq2:results]{DESeq2::results()}}. (default 2)}

\item{top.n}{whole number. Number of most highly regulated genes to label
with gene symbols on Volcano plots. (default 30).}

\item{useDingbats}{logical. When TRUE, some pdfs are made with useDingbats
set to TRUE, to reduce file sizes. (default FALSE)}

\item{print.all}{logical. When TRUE, output includes csv files of the whole
results set for each contrast (titled *_all), in addition to csv files
filtered for DEGs according to user-specified thresholds (titled *_DG),
which are required for \code{\link[=compare_reciprocal_contrasts]{compare_reciprocal_contrasts()}}. The *_all.csv
files may be useful for downstream analysis, such as GSEA.}
}
\value{
None - but creates pdf and csv files in the specified folder.
}
\description{
Runs \code{\link[=DESeq]{DESeq()}} for a \code{DESeqDataSet} with a design comparing multiple groups
(i.e. a factor with >2 levels). It uses each possible reference level in
turn, extracting all contrast results for each. The design may be
multi-factorial (e.g., accounting for batch effects) but
\code{\link[=run_DESeq_all_contrasts]{run_DESeq_all_contrasts()}} is not implemented to handle designs with
interaction terms. It is intended to facilitate the alternative strategy
suggested in the DESeq2 vignette, whereby interacting variables are combined
into a single new factor including many levels which are combinations of the
original variables.
}
\section{Further details}{
 Output includes pdf files with DESeq2 MA plots,
volcano plots and heatmaps of significant genes for each contrast in one
pdf file per reference level, and also a csv of the significant genes for
each contrast (according to the user-specified adj-pvalue and lfc
thresholds).
}

