% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeqAllContrasts.R
\name{run_DESeq_all_contrasts}
\alias{run_DESeq_all_contrasts}
\title{Run DESeq for all contrasts}
\usage{
run_DESeq_all_contrasts(
  dds,
  folder,
  mode = "lfcShrink",
  condition,
  rowname2symbol = NULL,
  heatmap.annotation.col = NULL,
  p.cutoff = 0.1,
  fc.cutoff = 2,
  top.n = 30,
  only.first.iteration = F
)
}
\arguments{
\item{dds}{Valid `DESeqDataSet` object with a simple design e.g. `design =
~condition`. If you wish to use a filter such as `edgeR` [filterByExpr()],
please do so prior to running this function. This package is intended to
help assess the usefulness of such filtering.}

\item{folder}{character string specifying the relative or absolute filepath
of the folder where you want the output files to go. This folder should
exist.}

\item{mode}{either `"noShrink"` or `"lfcShrink"`. Specifies whether to use
[results()] or [lfcShrink()] (default is `"lfcShrink"`).}

\item{condition}{character string related to the design formula used when
creating the `DESeqDataSet`. This should be a `colData` variable in the
`DESeqDataSet`; e.g. for `design=~group`, here enter `condition="group"`.
If you use a more complex design, results are not guaranteed. See also
`only.first.iteration`.}

\item{rowname2symbol}{data.frame or tibble with 2 columns titled `"rownames"`
and `"g_symbol"`. `"rownames"` should match `"rownames(dds)"`. This is used
for labeling the Volcano plot and annotating the tables of differentially
expressed genes}

\item{heatmap.annotation.col}{character vector specifying which `colData`
variables to use as groups in the heatmap output. If NULL, the heatmap will
use the variable specified in `condition` (default NULL)}

\item{p.cutoff}{numeric. The significance cut-off for p-adjusted values,
equivalent to parameter alpha in [DESeq2::results()]. (default 0.1)}

\item{fc.cutoff}{numeric. The log2 fold change significance cut-off; this is
used in the Volcano plots and when printing out significant genes. It is
NOT used to change the null hypothesis in [DESeq2::results()]. (default 2)}

\item{top.n}{whole number. Number of most highly regulated genes to label
with gene symbols on Volcano plots. (default 30).}

\item{only.first.iteration}{If True, the function will process all the
contrasts using the first level as the reference level, then exit. This
should produce 1 set of csvs for contrast DGEs and 1 pdf with MA and
Volcano plots. It may be helpful for looking at more complex designs, so
long as `condition` can still be specified as a `colData` variable. But no
promises. (default False)}
}
\value{
None - but creates pdf and csv files in the specified folder.
}
\description{
Runs [DESeq()] for a `DESeqDataSet` with a simple design. It uses each
possible reference level in turn, extracting all contrast results for each.
}
\section{Further details}{
 Output includes pdf files with DESeq2 MA plots,
  volcano plots and heatmaps of significant genes for each contrast in one
  pdf file per reference level, and also a csv of the significant genes for
  each contrast (according to the user-specified adj-pvalue and lfc
  thresholds). It is possible to run with more complex designs for the first
  reference level using the `only.first.iteration` parameter, but this has
  not been tested thoroughly and output should be interpreted with care.
}

